pyproject.toml
[project]
name = "commucraft-ai"
version = "0.1.0"
description = "AI-powered daily communication learning system using LangChain and Google Gemini API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "google-genai>=0.5.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=2.0.10",
    "google-auth>=2.30.0",
    "atlassian-python-api>=3.41.0",
    "sentence-transformers>=2.2.2",
]

[project.scripts]
commucraft-ai = "commucraft_ai.main:main"

[project.optional-dependencies]
dev = ["pytest>=8.4.2", "ruff>=0.15.7"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.15.7",
]
